% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_delete_comment_content}
\alias{codecommit_delete_comment_content}
\title{Deletes the content of a comment made on a change, file, or commit in a
repository}
\usage{
codecommit_delete_comment_content(commentId)
}
\arguments{
\item{commentId}{[required] The unique, system-generated ID of the comment. To get this ID, use
GetCommentsForComparedCommit or GetCommentsForPullRequest.}
}
\description{
Deletes the content of a comment made on a change, file, or commit in a
repository.
}
\section{Request syntax}{
\preformatted{codecommit$delete_comment_content(
  commentId = "string"
)
}
}

\keyword{internal}
