% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecommit_operations.R
\name{codecommit_delete_repository}
\alias{codecommit_delete_repository}
\title{Deletes a repository}
\usage{
codecommit_delete_repository(repositoryName)
}
\arguments{
\item{repositoryName}{[required] The name of the repository to delete.}
}
\description{
Deletes a repository. If a specified repository was already deleted, a
null repository ID will be returned.
}
\details{
Deleting a repository also deletes all associated objects and metadata.
After a repository is deleted, all future push calls to the deleted
repository will fail.
}
\section{Request syntax}{
\preformatted{codecommit$delete_repository(
  repositoryName = "string"
)
}
}

\keyword{internal}
